refactor(segmentation): extract editor JSX into presentational SegmentationEditorLayout
Pure presentational split of SegmentationEditor.tsx: the entire render
tree moves into a new SegmentationEditorLayout component that receives
every value/handler via an explicit typed Props interface. No hooks were
moved or reordered; no behavior change. The orchestrator keeps all state,
effects, refs, and logic, and its return now renders the layout inside the
existing SegmentationErrorBoundary.

Deliberately not wrapped in React.memo (an incomplete comparator is a
known recurring bug here). Hook-type imports use `import type` to keep the
hook modules out of the layout's runtime import graph.
